requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dj-database-url==2.1.0
h11==0.14.0
idna==3.10
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
gunicorn==21.2.0 